+++ logprep/ng/processor/decoder/processor.py
@@ -42,9 +42,9 @@ from typing import Callable
 
 from typing_extensions import override
 
+from logprep.ng.processor.field_manager.processor import FieldManager
 from logprep.processor.decoder.decoders import DECODERS, DecoderError
 from logprep.processor.decoder.rule import DecoderRule
-from logprep.processor.field_manager.processor import FieldManager
 from logprep.processor.field_manager.rule import FieldManagerRule
 from logprep.util.helper import FieldValue, add_fields_to
 from logprep.util.typing import is_list_of
@@ -79,5 +79,5 @@ class Decoder(FieldManager):
             raise DecoderError("can only decode string values")
         except DecoderError as error:
             add_fields_to(event, {"tags": rule.failure_tags}, merge_with_target=True)
-            self.result.errors.append(error)
+            self._event.errors.append(error)
             return []